Refactor notification handling and add notification body message
This commit refactors the notification handling in the `NotificationApi` class and adds a notification body message to be displayed in the notification. The `onReceivedMessage` function now checks if the user is the sender of the message and if so, does not show the notification. Additionally, a `notificationBody` function was added to the `group_chat_screen.dart` file to generate a notification body message based on the type of message being sent.