Refactor notification handling and add local notification click handler
This commit refactors the notification handling in the app to improve the user experience. It adds a local notification click handler that allows the user to navigate to the relevant screen when a notification is clicked. The commit also includes changes to the `GroupCubit` to keep track of whether the group screen is opened or not. Additionally, it includes changes to the `GroupHomeEntity` and `GroupNotificationEntity` to add a `screen` property that keeps track of the current screen of the group. Finally, the commit includes changes to the `UserMainCubit` to add a local notification click handler that navigates the user to the relevant screen when a notification is clicked.