Add saving of notifications to local data source
This commit adds the functionality to save notifications to the local data source when a message is received. It also includes the necessary changes to the `NotificationLocalDataSource` interface and implementation. The `saveNotification` method is used to save the notification to the local data source, and the `addNotifications` method is used to add the notification to the `notificationBox`. This change will help to keep track of notifications and display them to the user.