feat: Add localization to group chat screen and user notifications settings screen
This commit adds localization to the group chat screen and user notifications settings screen. It includes the addition of two new strings to the localization files: "Activate vibration" and "What action do you want to perform with this "{content}"?". The latter string is used in the EditMessageDialogWidget in the group chat screen. The former string is used in the NotificationItemWidget in the user notifications settings screen.