Refactor notification handling and add navigation to group screen
This commit refactors the notification handling in the `notification_services.dart` file to handle notifications in-app and adds a new function `handleNotificationInApp` to handle notifications when the app is opened. It also adds a new dependency `HomeLocalDataSource` to update the last activity of a group in the local database.

In addition, this commit adds a new feature to navigate to the group screen when a notification is tapped. It also adds a new boolean variable `isNotificationScreenOpened` to track whether the notification screen is opened or not.

Lastly, this commit updates the `UserNotificationCubit` to insert new notifications when a new activity is received and the notification screen is opened. It also updates the `UserNotificationScreen` to navigate to the group screen when a notification is tapped.