Add group media screen and route
Added a new screen and route for displaying media and documents for a group. The screen is implemented as a `CustomScrollView` with a `SliverAppBar` displaying the group name. The route is added to the `AppRoute` and `GroupRoutes` classes. The `MediaDocsTileGroupDetails` widget is updated to navigate to the new screen when tapped, passing in the `ListResult` object.